requirements.txt
# Web Framework & Extensions
Flask>=2.3.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.0

# Data Processing & Math
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# Machine Learning & Model Loading
scikit-learn>=1.3.0    # 載入 joblib 模型通常需要 sklearn 環境
joblib>=1.3.0
shap>=0.42.1           # 雖然程式碼讀取圖片，但訓練端通常需要這個

# AI / LLM Integration
google-genai>=0.2.0           # <-- 改成這一行 (支援 genai.Client)

# Server (Production use)
gunicorn>=21.2.0       # 部署時建議使用